requirements.txt
arrow==0.10.0
binaryornot==0.4.0
chardet==2.3.0
click==6.7
cookiecutter==1.5.1
Django==1.10.5
django-filer==1.2.6rc2
django-model-utils==2.6.1
django-mptt==0.8.7
django-polymorphic==1.0.2
easy-thumbnails==2.3
future==0.16.0
Jinja2==2.9.5
jinja2-time==0.2.0
MarkupSafe==0.23
olefile==0.44
Pillow==4.0.0
poyo==0.4.0
psycopg2==2.6.2
python-dateutil==2.6.0
six==1.10.0
Unidecode==0.4.20
whichcraft==0.4.0
